Refactor placing of actions.
