+++ app/actions/InvoiceGridActions.js
@@ -1,0 +1,19 @@
+const GET_INVOICES_REQUEST = "GET_INVOICES_REQUEST";
+export function getInvoicesRequest() {
+  return {
+    type: GET_INVOICES_REQUEST
+  }
+}
+
+export function getInvoicesForGrid() {
+  return (dispatch) => {
+    dispatch(getInvoicesRequest());
+    fetch('http://192.168.99.100:7004/api/invoices/', {
+      method: 'GET',
+      headers: new Headers({
+        'Content-Type': 'application/json'
+      })
+    }).then(res => res.json())
+      .then(x => console.log(x));
+  };
+}
