+++ app/containers/Fields/ChooseDate.js
@@ -5,7 +5,7 @@ import * as actions from '../../actions';
 const mapDispatchToProps = (dispatch, ownProps) => {
   return {
     onChange: (event, date) => {
-      dispatch(actions.updateUploadForm('date', date.getTime()))
+      dispatch(actions.uploadInvoice.updateUploadForm('date', date.getTime()))
     }
   }
 };
